backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
scikit-learn
pandas
numpy
python-dotenv
requests
python-jose[cryptography]
passlib[bcrypt]
PyJWT
python-multipart
aiofiles
